VALIDATION/requirements.txt
azure-storage-blob>=12.21.0
ujson>=5.10.0
